requirements.txt
python-telegram-bot==20.7
pillow>=9.0
aiohttp>=3.8
python-dotenv>=1.0.0
requests>=2.28.0
pytz>=2023.3
pymongo==4.8.0
pyfiglet>=0.8
motor>=3.5.1
httpx==0.25.2  # Changed from 0.27.0 to match python-telegram-botaiohttp>=3.8
